Extract app member people resolution and enforce user UUIDs.
Person lookup, workspace directory resolution, and subject-id validation now live in authorization_app_member_people, and user subjects must be canonical UUIDs rather than opaque or email-shaped ids.
